% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_genomic_features.r
\name{add_genomic_features,Vplots,GRanges-method}
\alias{add_genomic_features,Vplots,GRanges-method}
\title{add_genomic_features}
\usage{
\S4method{add_genomic_features}{Vplots,GRanges}(x, y, x_name = NULL, y_name = NULL, ...)
}
\arguments{
\item{x}{a Vplots object}

\item{y}{a GRanges object}

\item{x_name}{rowData(x) field for storing the features}

\item{y_name}{mcols(y) field for accessing the features}

\item{...}{other parameters}
}
\value{
a Vplots object
}
\description{
Add genomic features
}
\author{
Wuming Gong (gongx030@umn.edu)
}
