% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.r
\name{add_genomic_features}
\alias{add_genomic_features}
\title{add_genomic_features}
\usage{
add_genomic_features(x, y, ...)
}
\arguments{
\item{x}{a data object that include the Vplot information}

\item{y}{a data object}

\item{...}{Other arguments}
}
\description{
The generic function of add_genomic_features
}
