% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_schedule.r
\name{beta_schedule}
\alias{beta_schedule}
\title{beta_schedule}
\usage{
beta_schedule(method = "constant", beta0 = 5e-05, epochs, ...)
}
\arguments{
\item{method}{Scheduling method: 'constant', 'monotonic' or 'cyclical_annealing'}

\item{beta0}{initial beta value}

\item{epochs}{Number of total epochs}

\item{...}{Additional arguments}
}
\description{
Schedule for beta during the VAE training
}
