% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{block_center}
\alias{block_center}
\title{block_center}
\usage{
block_center(x)
}
\arguments{
\item{x}{length of the sequence}
}
\value{
a binary vector indicating the center of the sequence with length x
}
\description{
Label the center of a x-length sequence
}
