% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.r
\name{call_nucleosomes}
\alias{call_nucleosomes}
\title{call_nucleosomes}
\usage{
call_nucleosomes(model, x, ...)
}
\arguments{
\item{model}{a Model}

\item{x}{a data object that include the Vplot information}

\item{...}{Other arguments}
}
\description{
The generic function of call_nucleosomes
}
