% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbind.r
\name{cbind,Vplots-method}
\alias{cbind,Vplots-method}
\title{Binding Vplots objects column-wise}
\usage{
\S4method{cbind}{Vplots}(..., deparse.level = 1)
}
\arguments{
\item{...}{Vplots objects}

\item{deparse.level}{integer controlling the construction of labels in the case of non-matrix-like arguments}
}
\value{
a Vplots object
}
\description{
Binding Vplots objects column-wise
}
