% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.r
\name{count_reads}
\alias{count_reads}
\title{count_reads}
\usage{
count_reads(x, filenames, genome, ...)
}
\arguments{
\item{x}{a genomic object}

\item{filenames}{BAM file names}

\item{genome}{a genome object}

\item{...}{Other arguments}
}
\description{
The generic function of count_reads
}
