% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{cut_data}
\alias{cut_data}
\title{cut data}
\usage{
cut_data(n, batch_size)
}
\arguments{
\item{n}{Length of the sequence}

\item{batch_size}{Batch size}
}
\value{
a list of sequence indices, where the length of the list is the number of batches.
}
\description{
Cut a sequence into small batches
}
\author{
Wuming Gong (gongx030@umn.edu)
}
