% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{downsample_vplot}
\alias{downsample_vplot}
\title{downsample_vplot}
\usage{
downsample_vplot(x, num_reads = 1L)
}
\arguments{
\item{x}{a tensorflow.tensor object of Vplots}

\item{num_reads}{The target number of reads (default: 1L)}
}
\value{
down-sampled Vplots in tensorflow.tensor object
}
\description{
Downsample a dense V-plot to a saprse V-plot, where the remaining number of reads is specified as `num_reads`.
}
