% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvae.r
\name{fit,mVaeModel,tf_dataset-method}
\alias{fit,mVaeModel,tf_dataset-method}
\title{fit}
\usage{
\S4method{fit}{mVaeModel,tf_dataset}(
  model,
  x,
  batch_size = 128L,
  epochs = 100L,
  learning_rate = 0.001,
  compile = TRUE,
  beta = 5e-05
)
}
\arguments{
\item{model}{a mVaeModel object, initialized by `new('mVaeModel', model = mVaeModel(...))`}

\item{x}{a tf_dataset object}

\item{batch_size}{Batch size (default: 128L)}

\item{epochs}{Number of training epochs (default: 100L)}

\item{learning_rate}{Learning rate (default: 1e-3)}

\item{compile}{Whether or not compile the tensorflow model (default: TRUE)}

\item{beta}{Beta sequences (default: 5e-5)}
}
\value{
a mVaeModel
}
\description{
Fit a mVaeModel
}
