% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vae.r
\name{fit,VaeModel,VplotsList-method}
\alias{fit,VaeModel,VplotsList-method}
\title{fit}
\usage{
\S4method{fit}{VaeModel,VplotsList}(model, x, ...)
}
\arguments{
\item{model}{a VaeModel object, initialized by `new('VaeModel', model = VaeModel(...))`}

\item{x}{a VplotsList object}

\item{...}{Additional arguments to fit('VaeModel`, 'list', ...)}
}
\value{
a VaeModel
}
\description{
Fit a VaeModel
}
