% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{get_fragment_size_per_batch}
\alias{get_fragment_size_per_batch}
\title{get_fragment_size_per_batch}
\usage{
get_fragment_size_per_batch(x)
}
\arguments{
\item{x}{a Vplot object}
}
\description{
Get the aggregated fragment size distribution per batch
}
