% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nucleosome_diff.r
\name{get_nucleosome_diff}
\alias{get_nucleosome_diff}
\title{get_nucleosome_diff}
\usage{
get_nucleosome_diff(x, y, fragment_size_threshold = 150)
}
\arguments{
\item{x}{a Vplots object}

\item{y}{a Vplots object}

\item{fragment_size_threshold}{Fragment size threshold where the any fragment size above is considered as nucleosome reads and below is considered as nucleosome free reads}
}
\description{
Get nucleosome difference between two V-plot objects
}
\author{
Wuming Gong (gongx030@umn.edu)
}
