% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{get_nucleosome_score}
\alias{get_nucleosome_score}
\title{get_nucleosome_score}
\usage{
get_nucleosome_score(x, beta0 = -8.2399, beta1 = 13.19315)
}
\arguments{
\item{x}{a tensorflow.tensor object}

\item{beta0}{Scale factor for calculating the nucleosome score}

\item{beta1}{Offset factor for calculating the nucleosome score}
}
\value{
a tensorflow.tensor object of the calculated nucleosome score
}
\description{
Calcualting the nucleosome score (eta) from Vplots
}
\author{
Wuming Gong (gongx030@umn.edu)
}
