% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragment_size.r
\name{getFragmentSizeMatrix}
\alias{getFragmentSizeMatrix}
\title{getFragmentSizeMatrix}
\usage{
getFragmentSizeMatrix(filenames, which, genome, window_size = 2000,
  bin_size = 20, step_size = 20, fragment_size_range = c(0, 500),
  fragment_size_interval = 10, min_reads_per_window = 20)
}
\description{
getFragmentSizeMatrix
}
