% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{kernel_smoothing_1d}
\alias{kernel_smoothing_1d}
\title{kernel_smoothing_1d}
\usage{
kernel_smoothing_1d(x, size, sigma)
}
\arguments{
\item{x}{a 2D tensor}

\item{size}{kernel size}

\item{sigma}{standard deviation of smoothing kernel}
}
\value{
a smoothed 2D tensor
}
\description{
1D kernel smoothing
}
