% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_model.r
\name{load_model,VaeModel,character-method}
\alias{load_model,VaeModel,character-method}
\title{load_model}
\usage{
\S4method{load_model}{VaeModel,character}(model, filename)
}
\arguments{
\item{model}{an empty VaeModel}

\item{filename}{Filename of the pretrained model}
}
\description{
Load a pretrianed model
}
\author{
Wuming Gong (gongx030@umn.edu)
}
