% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvae.r
\name{predict,mVaeModel,Vplots-method}
\alias{predict,mVaeModel,Vplots-method}
\title{predict}
\usage{
\S4method{predict}{mVaeModel,Vplots}(model, x, batch_size = 256L, vplots = TRUE, ...)
}
\arguments{
\item{model}{a trained mVaeModel object}

\item{x}{a Vplots object}

\item{batch_size}{Batch size (default: 256L)}

\item{vplots}{Whether or not return predicted Vplots as assays(x)$predicted_counts}

\item{...}{Additional arguments}
}
\value{
a Vplots object
}
\description{
Predict counts
}
\author{
Wuming Gong (gongx030@umn.edu)
}
