% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vae.r
\name{predict,VaeModel,VplotsList-method}
\alias{predict,VaeModel,VplotsList-method}
\title{predict}
\usage{
\S4method{predict}{VaeModel,VplotsList}(
  model,
  x,
  batch_size = 256L,
  vplots = FALSE,
  nucleosome = TRUE,
  fragment_size_threshold = 150L,
  ...
)
}
\arguments{
\item{model}{a trained VaeModel object}

\item{x}{a VplotsList object}

\item{batch_size}{Batch size (default: 256L)}

\item{vplots}{Whether or not return predicted Vplots as assays(x)$predicted_counts}

\item{nucleosome}{Whether or not return predicted nucleosome as rowData(x)$predicted_nucleosome}

\item{fragment_size_threshold}{Fragment size threshold for nucleosome reads (default: 150L)}

\item{...}{Additional arguments}
}
\value{
a Vplots object
}
\description{
Predict counts
}
\author{
Wuming Gong (gongx030@umn.edu)
}
