% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvae.r
\name{predict_fragment_size,mVaeModel,Vplots-method}
\alias{predict_fragment_size,mVaeModel,Vplots-method}
\title{predicted_fragment_size}
\usage{
\S4method{predict_fragment_size}{mVaeModel,Vplots}(model, x, batch_size = 256L, width = 100L, ...)
}
\arguments{
\item{model}{a trained mVaeModel object}

\item{x}{a Vplots object}

\item{batch_size}{Batch size (default: 256L)}

\item{width}{Central width (in bp) for computing fragment size disstribution (default: 100L)}

\item{...}{Other arguments passed to prepare_data.}
}
\value{
a Vplots object
}
\description{
Predict fragment size at the center of the Vplot
}
\author{
Wuming Gong (gongx030@umn.edu)
}
