% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.r
\name{predict_fragment_size}
\alias{predict_fragment_size}
\title{predict_fragment_size}
\usage{
predict_fragment_size(model, x, ...)
}
\arguments{
\item{model}{a model object}

\item{x}{a data object that include the Vplot information}

\item{...}{Other arguments}
}
\description{
The generic function of predict_fragment_size
}
