% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.r
\name{predict_vplots}
\alias{predict_vplots}
\title{predict_vplots}
\usage{
predict_vplots(model, x, ...)
}
\arguments{
\item{model}{a model object}

\item{x}{a data object that include the Vplot information}

\item{...}{Other arguments}
}
\description{
The generic function of predict_vplots
}
