% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvae.r
\name{prepare_data,mVaeModel,Vplots-method}
\alias{prepare_data,mVaeModel,Vplots-method}
\title{prepare_data}
\usage{
\S4method{prepare_data}{mVaeModel,Vplots}(model, x, ...)
}
\arguments{
\item{model}{a mVaeModel object, initialized by `new('mVaeModel', model = mVaeModel(...))`}

\item{x}{a Vplots object}

\item{...}{Other arguments}
}
\value{
a list that include `vplots` and `batch`
}
\description{
Prepare dataset for training and a V-plot model
}
\author{
Wuming Gong (gongx030@umn.edu)
}
