% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vae.r
\name{prepare_data,VaeModel,Vplots-method}
\alias{prepare_data,VaeModel,Vplots-method}
\title{prepare_data}
\usage{
\S4method{prepare_data}{VaeModel,Vplots}(model, x, ...)
}
\arguments{
\item{model}{a VaeModel object, initialized by `new('VaeModel', model = VaeModel(...))`}

\item{x}{a Vplots object}

\item{...}{Other arguments}
}
\value{
a list that include `vplots` and `batch`
}
\description{
Prepare dataset for training and a V-plot model
}
