% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vae.r
\name{prepare_data,VaeModel,VplotsList-method}
\alias{prepare_data,VaeModel,VplotsList-method}
\title{prepare_data}
\usage{
\S4method{prepare_data}{VaeModel,VplotsList}(model, x, ...)
}
\arguments{
\item{model}{a VaeModel object, initialized by `VaeModel()`}

\item{x}{a VplotsList object}

\item{...}{Other arguments}
}
\value{
a list that include `vplots` and `batch`
}
\description{
Prepare dataset for training and a V-plot model
}
