% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.r
\name{prepare_data}
\alias{prepare_data}
\title{prepare_data}
\usage{
prepare_data(model, x, ...)
}
\arguments{
\item{model}{a model object}

\item{x}{a data object that include the Vplot information}

\item{...}{Other arguments}
}
\description{
The generic function of prepare_data
}
