% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_aggregated_vplot.r
\name{read_aggregated_vplot,GRangesList,character,BSgenome-method}
\alias{read_aggregated_vplot,GRangesList,character,BSgenome-method}
\title{Read aggregated V-plot from a list of genomic ranges (GRangeList)}
\usage{
\S4method{read_aggregated_vplot}{GRangesList,character,BSgenome}(
  x,
  filenames,
  genome,
  bin_size = 5L,
  fragment_size_range = c(80, 320),
  fragment_size_interval = 5L
)
}
\arguments{
\item{x}{a GRangeList object that define a list of genomic region sets.}

\item{filenames}{BAM file names}

\item{genome}{a BSgenome object}

\item{bin_size}{The bin size in base pairs (default: 5L)}

\item{fragment_size_range}{The range of the PE reads fragment sizes (default: c(80L, 320L))}

\item{fragment_size_interval}{Fragment size interval (default: 5L)}
}
\value{
a SummarizedVplots object
}
\description{
This function is to quickly read aggregated V-plot given a list of GRanges (e.g. motif binding sites) for 
scATAC-seq.
}
\author{
Wuming Gong (gongx030@umn.edu)
}
