% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bam.r
\name{read_bam,character,GRanges,BSgenome-method}
\alias{read_bam,character,GRanges,BSgenome-method}
\title{Read BAM file}
\usage{
\S4method{read_bam}{character,GRanges,BSgenome}(filename, peaks, genome)
}
\arguments{
\item{filename}{BAM file names}

\item{peaks}{a GRange object that define a set of genomic regions.}

\item{genome}{a BSgenome for genome information}
}
\description{
Read the aligned paired-end reads over the given peaks.
}
\author{
Wuming Gong (gongx030@umn.edu)
}
