% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bam.r
\name{read_bam,character,missing,BSgenome-method}
\alias{read_bam,character,missing,BSgenome-method}
\title{Read BAM file}
\usage{
\S4method{read_bam}{character,missing,BSgenome}(filename, peaks, genome)
}
\arguments{
\item{filename}{BAM file names}

\item{peaks}{missing}

\item{genome}{a BSgenome for genome information}
}
\description{
Read all aligned paired-end reads
}
\author{
Wuming Gong (gongx030@umn.edu)
}
