% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.r
\name{read_bam}
\alias{read_bam}
\title{read_bam}
\usage{
read_bam(filename, peaks, genome, ...)
}
\arguments{
\item{filename}{BAM file name}

\item{peaks}{a peak object}

\item{genome}{a genome object}

\item{...}{Other arguments}
}
\description{
The generic function of read_bam
}
