% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fragment_size_profile.r
\name{read_fragment_size_profile,GRanges,character,BSgenome-method}
\alias{read_fragment_size_profile,GRanges,character,BSgenome-method}
\title{Read fragment size profile}
\usage{
\S4method{read_fragment_size_profile}{GRanges,character,BSgenome}(
  x,
  filename,
  genome,
  fragment_size_range = c(0, 320L),
  fragment_size_interval = 10L
)
}
\arguments{
\item{x}{a GRange object that define a set of genomic regions.}

\item{filename}{BAM file name}

\item{genome}{a BSgenome object}

\item{fragment_size_range}{The range of the PE reads fragment sizes that are used for 
constructing Vplot (default: c(0L, 320L))}

\item{fragment_size_interval}{Fragment size interval (default: 10L)}
}
\description{
Read fragment size profile from BAM files
}
\author{
Wuming Gong (gongx030@umn.edu)
}
