% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.r
\name{read_fragment_size_profile}
\alias{read_fragment_size_profile}
\title{read_fragment_size_profile}
\usage{
read_fragment_size_profile(x, filename, genome, ...)
}
\arguments{
\item{x}{a genomic object}

\item{filename}{BAM file name}

\item{genome}{a genome object}

\item{...}{Other arguments}
}
\description{
The generic function of read_fragment_size_profile
}
