% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vplot.r
\name{read_vplot,GRangesList,character,BSgenome-method}
\alias{read_vplot,GRangesList,character,BSgenome-method}
\title{Read the V-plot}
\usage{
\S4method{read_vplot}{GRangesList,character,BSgenome}(x, filenames, genome, ...)
}
\arguments{
\item{x}{a GRangesList object that define a set of genomic regions.}

\item{filenames}{BAM file names}

\item{genome}{a BSgenome object}

\item{...}{Additional arguments for read_vplot('GRanges', 'character', 'BSgenome', ...)}
}
\value{
a VplotsList object
}
\description{
Read the V-plot from BAM files within a set of genomic regions
}
\author{
Wuming Gong (gongx030@umn.edu)
}
