% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.r
\name{read_vplot}
\alias{read_vplot}
\title{read_vplot}
\usage{
read_vplot(x, filenames, genome, ...)
}
\arguments{
\item{x}{a genomic object}

\item{filenames}{BAM file name}

\item{genome}{a genome object}

\item{...}{Other arguments}
}
\description{
The generic function of read_vplot
}
