% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vplot.r
\name{read_vplot_core}
\alias{read_vplot_core}
\title{Read the V-plot}
\usage{
read_vplot_core(
  x,
  filename,
  genome,
  bin_size = 5L,
  fragment_size_range = c(0L, 320L),
  fragment_size_interval = 10L
)
}
\arguments{
\item{x}{a GRange object that define a set of genomic regions.}

\item{filename}{BAM file names}

\item{genome}{a BSgenome object}

\item{bin_size}{The bin size (default: 5L)}

\item{fragment_size_range}{The range of the PE reads fragment sizes that are used for 
constructing Vplot (default: c(0L, 320L))}

\item{fragment_size_interval}{Fragment size interval (default: 10L)}
}
\description{
Read the V-plot from BAM files within a set of genomic regions
}
\author{
Wuming Gong (gongx030@umn.edu)
}
