% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragment_size.r
\name{readFragmentSize}
\alias{readFragmentSize}
\title{readFragmentSize}
\usage{
readFragmentSize(filename, bins, fragment_size_range = c(50, 1000),
  fragment_size_interval = 10)
}
\description{
readFragmentSize
}
