% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resize.r
\name{resize,Vplots-method}
\alias{resize,Vplots-method}
\title{Resize Vplots}
\usage{
\S4method{resize}{Vplots}(x, width, fix = "center")
}
\arguments{
\item{x}{an input Vplots object}

\item{width}{Width of the sub-window}

\item{fix}{a character vector containing the values "start", "end", or "center" denoting what to use as an anchor for each element in x.}
}
\value{
a new Vplots object
}
\description{
Resize a Vplots object
}
