% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.r
\name{results,VaeModel,Vplots-method}
\alias{results,VaeModel,Vplots-method}
\title{results}
\usage{
\S4method{results}{VaeModel,Vplots}(model, x, type = NULL, contrast = NULL, ...)
}
\arguments{
\item{model}{a trained VaeModel object}

\item{x}{a Vplots object}

\item{type}{Tesing type ('phase', 'nucleosome', or 'vplots')}

\item{contrast}{this argument species a character vector with exactly three elements: the name of a factor in the design formula, the name of the numerator level for the fold change, and the name of the denominator level for the fold change (simplest case)}

\item{...}{Other arguments}
}
\value{
a GRanges object
}
\description{
Test the difference between two Vplots
}
\author{
Wuming Gong (gongx030@umn.edu)
}
