% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.r
\name{results_phase}
\alias{results_phase}
\title{results_phase}
\usage{
results_phase(
  model,
  x,
  contrast,
  width = 100L,
  repeats = 100L,
  batch_size = 16L
)
}
\arguments{
\item{model}{a trained VaeModel object}

\item{x}{a Vplots object}

\item{contrast}{this argument species a character vector with exactly three elements: the name of a factor in the design formula, the name of the numerator level for the fold change, and the name of the denominator level for the fold change (simplest case)}

\item{width}{Width of the centeral regions to be considered (default: 100L)}

\item{repeats}{Number of repeated sampling (default: 100L)}

\item{batch_size}{Batch size (default: 16L)}
}
\value{
a GRanges object
}
\description{
Test the difference of the phase between two Vplots
}
\author{
Wuming Gong (gongx030@umn.edu)
}
