% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.r
\name{results_vplots}
\alias{results_vplots}
\title{results_vplots}
\usage{
results_vplots(model, x, contrast, batch_size = 128L)
}
\arguments{
\item{model}{a trained VaeModel object}

\item{x}{a Vplots object}

\item{contrast}{this argument species a character vector with exactly three elements: the name of a factor in the design formula, the name of the numerator level for the fold change, and the name of the denominator level for the fold change (simplest case)}

\item{batch_size}{Batch size (default: 128L)}
}
\value{
a GRanges object
}
\description{
Test the difference of between two Vplots
}
\author{
Wuming Gong (gongx030@umn.edu)
}
