% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale01.r
\name{scale01,dgCMatrix-method}
\alias{scale01,dgCMatrix-method}
\title{scale01}
\usage{
\S4method{scale01}{dgCMatrix}(x)
}
\arguments{
\item{x}{a dgCMatrix object}
}
\value{
a scaled dgCMatrix object
}
\description{
Scale the column a matrix to [0, 1]
}
