% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale01.r
\name{scale01,matrix-method}
\alias{scale01,matrix-method}
\title{scale01}
\usage{
\S4method{scale01}{matrix}(x)
}
\arguments{
\item{x}{matrix object}
}
\value{
a scaled matrix object
}
\description{
Scale the column a matrix to [0, 1]
}
