% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale01.r
\name{scale01,numeric-method}
\alias{scale01,numeric-method}
\title{scale01}
\usage{
\S4method{scale01}{numeric}(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a scaled numeric vector
}
\description{
Scale the column a matrix to [0, 1]
}
