% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale01.r
\name{scale01,tensorflow.tensor-method}
\alias{scale01,tensorflow.tensor-method}
\title{scale01}
\usage{
\S4method{scale01}{tensorflow.tensor}(x)
}
\arguments{
\item{x}{a tensorflow.tensor object}
}
\value{
a scaled tensorflow.tensor object
}
\description{
Scale the column a matrix to [0, 1]
}
