% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{scale_vplot}
\alias{scale_vplot}
\title{scale_vplot}
\usage{
scale_vplot(x)
}
\arguments{
\item{x}{Vplots in tensorflow.tensor object ([batch, height, width, 1])}
}
\value{
scaled Vplots in tensorflow.tensor object ([batch, height, width, 1])
}
\description{
Scale the Vplot so that the sum of reads at each position is one.
}
\author{
Wuming Gong (gongx030@umn.edu)
}
