% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seatac.r
\name{seatac}
\alias{seatac}
\title{seatac}
\usage{
seatac(filenames, which = NULL, genome, latent_dim = 10,
  num_states = 3, width = 200, expand = 2000,
  fragment_size_range = c(70, 500), fragment_size_interval = 10,
  min_reads_per_bin = 20, epochs = 5, batch_size = 256)
}
\description{
Integrating multiple sources of temporal scRNA-seq data by neighborhood component analysis
}
\author{
Wuming Gong
}
