% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.r
\name{show,VplotsList-method}
\alias{show,VplotsList-method}
\title{show-VplotsList}
\usage{
\S4method{show}{VplotsList}(object)
}
\arguments{
\item{object}{a VplotsList object}
}
\description{
Print a VplotsList object
}
