% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slidingWindows.r
\name{slidingWindows,Vplots-method}
\alias{slidingWindows,Vplots-method}
\title{slidingWindows}
\usage{
\S4method{slidingWindows}{Vplots}(x, width = 640L, step = 20L, batch_size = 4096L)
}
\arguments{
\item{x}{a Vplots object}

\item{width}{The width of the sliding window (default: 640L)}

\item{step}{The step for each sliding window (default: 20L)}

\item{batch_size}{The batch size of extraction operation (default: 4096L)}
}
\value{
a new Vplot object which window_size is the specified width. The rowData and colData 
in the input Vplots object will be missing in the new Vplot object.
}
\description{
Get the sliding V-plots
}
