% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting.r
\name{[,Vplots,ANY,ANY,ANY-method}
\alias{[,Vplots,ANY,ANY,ANY-method}
\title{Subsetting a Vplots object}
\usage{
\S4method{[}{Vplots,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{a Vplots object}

\item{i}{row-wise indexing}

\item{j}{column-wise indexing}

\item{...}{Additional arguments}

\item{drop}{Whether or not drop dimension}
}
\description{
Subsetting a Vplots object
}
