% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting.r
\name{[<-,Vplots,ANY,ANY,Vplots-method}
\alias{[<-,Vplots,ANY,ANY,Vplots-method}
\title{Subsetting a Vplots object}
\usage{
\S4method{[}{Vplots,ANY,ANY,Vplots}(x, i, j, ...) <- value
}
\arguments{
\item{x}{a Vplots object}

\item{i}{row-wise indexing}

\item{j}{column-wise indexing}

\item{...}{Additional arguments}

\item{value}{The replacement value}
}
\description{
Subsetting a Vplots object
}
