% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.r
\name{test_accessibility}
\alias{test_accessibility}
\title{test_accessibility}
\usage{
test_accessibility(model, x, ...)
}
\arguments{
\item{model}{a model object}

\item{x}{a data object that include the Vplot information}

\item{...}{Other arguments}
}
\description{
The generic function of test_accessibility
}
