% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{until_out_of_range2}
\alias{until_out_of_range2}
\title{Execute code that traverses a dataset until an out of range condition occurs}
\usage{
until_out_of_range2(expr)
}
\arguments{
\item{expr}{an expression}
}
\description{
https://www.rdocumentation.org/packages/tfdatasets/versions/2.7.0/topics/until_out_of_range
As of 5/17/2022, `until_out_of_range` from the tfdatasets failed on colab, due to the following error:
Error: tensorflow.python.framework.errors_impl.OutOfRangeError: End of sequence [Op:IteratorGetNext]
}
