% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vae.r
\name{validate,VaeModel,Vplots-method}
\alias{validate,VaeModel,Vplots-method}
\title{validate}
\usage{
\S4method{validate}{VaeModel,Vplots}(model, x)
}
\arguments{
\item{model}{a VaeModel object}

\item{x}{a Vplots object}

\item{...}{Other arguments passed to prepare_data.}
}
\value{
a logical value of whether the data are valid.
}
\description{
Validate the input data for current model
}
