% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vae.r
\name{validate,VaeModel,VplotsList-method}
\alias{validate,VaeModel,VplotsList-method}
\title{validate}
\usage{
\S4method{validate}{VaeModel,VplotsList}(model, x)
}
\arguments{
\item{model}{a VaeModel object}

\item{x}{a VplotsList object}
}
\value{
a logical value of whether the data are valid.
}
\description{
Validate the input data for current model
}
