% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.r
\name{validate}
\alias{validate}
\title{validate}
\usage{
validate(model, x)
}
\arguments{
\item{model}{a Model}

\item{x}{a data object that include the Vplot information}
}
\description{
The generic function of validate
}
