% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vplot.r
\name{vplot,Vplots-method}
\alias{vplot,Vplots-method}
\title{vplot}
\usage{
\S4method{vplot}{Vplots}(x, field = "counts", ncol = 2L)
}
\arguments{
\item{x}{a Vplots object}

\item{field}{The assays field that is used for Vplots (default: 'counts')}

\item{ncol}{number of columns (default: 2L)}
}
\description{
Plot a Vplot object
}
