% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vplot.r
\name{vplot_core}
\alias{vplot_core}
\title{vplot_core}
\usage{
vplot_core(x, field, ncol = 2)
}
\arguments{
\item{x}{a Vplots object}

\item{field}{The assays field(s) that is used for Vplots (default: 'counts')}

\item{ncol}{number of columns (default: 2L)}
}
\description{
Plot a Vplot object
}
