% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vae.r
\name{load_model,VaeModel-method}
\alias{load_model,VaeModel-method}
\title{load_model}
\usage{
\S4method{load_model}{VaeModel}(model, filename = "character", ...)
}
\arguments{
\item{model}{a VaeModel object}
}
\value{
a Vplots object
}
\description{
Load a pretrained VaeModel
}
\author{
Wuming Gong (gongx030@umn.edu)
}
