% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.r
\name{load_model}
\alias{load_model}
\title{load_model}
\usage{
load_model(model, filename, ...)
}
\arguments{
\item{model}{a Model}

\item{filename}{Filename of the pretrained model}

\item{...}{Other arguments}
}
\description{
The generic function of load_model
}
