% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vplot.r
\name{read_vplot,GRangesList,character,BSgenome-method}
\alias{read_vplot,GRangesList,character,BSgenome-method}
\title{Read the V-plot}
\usage{
\S4method{read_vplot}{GRangesList,character,BSgenome}(
  x,
  filenames,
  genome,
  bin_size = 5L,
  fragment_size_range = c(0, 320L),
  fragment_size_interval = 10L,
  ignore_strand = TRUE
)
}
\arguments{
\item{x}{a GRangesList object that define a set of genomic regions.}

\item{filenames}{BAM file names}

\item{genome}{a BSgenome object}

\item{bin_size}{The bin size (default: 5L)}

\item{fragment_size_range}{The range of the PE reads fragment sizes that are used for 
constructing Vplot (default: c(0, 320L))}

\item{fragment_size_interval}{Fragment size interval (default: 10L)}

\item{ignore_strand}{whether ignore the strand of the V-plot (default: TRUE)}
}
\description{
Read the V-plot from BAM files within a set of genomic regions
}
\author{
Wuming Gong (gongx030@umn.edu)
}
