% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.r
\name{results_vplots}
\alias{results_vplots}
\title{results_vplots}
\usage{
results_vplots(x, field, treatment, control, width = 100L)
}
\arguments{
\item{x}{a Vplots object}

\item{field}{The field in x@dimdata[['sample']] that defines the sample groups}

\item{treatment}{the name of the numerator level for the fold change}

\item{control}{the name of the denominator level for the fold change}

\item{width}{Width of the centeral regions to be considered (default: 100L)}
}
\value{
a GRanges object
}
\description{
Test the difference of Vplots by a Chi-squared test
}
