% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seatac.r
\name{seatac}
\alias{seatac}
\title{seatac}
\usage{
seatac(x, latent_dim = 2, window_size = 640,
  min_reads_per_window = 20, epochs = 50, batch_size = 128,
  steps_per_epoch = 20, sequence_dim = 32, n_components = 15,
  core_width = 100, batch_size_prediction = 64, genome, ...)
}
\description{
seatac
}
\author{
Wuming Gong
}
