% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seatac.r
\name{seatac}
\alias{seatac}
\title{seatac}
\usage{
seatac(filenames, time_points, genome, latent_dims = 10,
  num_states = 2, width = 200, expand = 2000, tilewidth = 1e+07,
  fragment_size_range = c(70, 750), fragment_size_interval = 10,
  min_reads_per_bin = 20, epochs = 5)
}
\description{
Integrating multiple sources of temporal scRNA-seq data by neighborhood component analysis
}
\author{
Wuming Gong
}
