% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seatac.r
\name{seatac}
\alias{seatac}
\title{seatac}
\usage{
seatac(x, type = "vae", latent_dim = 2L, hidden_dim = 8L,
  epochs = 50, batch_size = 256, steps_per_epoch = 10, ...)
}
\description{
seatac
}
\author{
Wuming Gong
}
