% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seatac.r
\name{seatac}
\alias{seatac}
\title{seatac}
\usage{
seatac(filenames, which = NULL, genome, window_size = 2000,
  bin_size = 20, fragment_size_range = c(70, 500),
  fragment_size_interval = 10, epochs = 5, batch_size = 256,
  gpu = TRUE)
}
\description{
Integrating multiple sources of temporal scRNA-seq data by neighborhood component analysis
}
\author{
Wuming Gong
}
