% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seatac.r
\name{seatac}
\alias{seatac}
\title{seatac}
\usage{
seatac(x, latent_dim = 2, epochs = 50, batch_size = 256,
  min_reads_per_window = 5, sequence_dim = 16L, type = "cvae")
}
\description{
seatac
}
\author{
Wuming Gong
}
