% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seatac.r
\name{seatac}
\alias{seatac}
\title{seatac}
\usage{
seatac(
  x,
  latent_dim = 2L,
  epochs = 50,
  batch_size = 128,
  steps_per_epoch = 50,
  type = "vae_baseline",
  ...
)
}
\description{
seatac
}
\author{
Wuming Gong
}
