% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seatac.r
\name{seatac}
\alias{seatac}
\title{seatac}
\usage{
seatac(x, latent_dim = 2, epochs = 50, batch_size = 256)
}
\description{
Integrating multiple sources of temporal scRNA-seq data by neighborhood component analysis
}
\author{
Wuming Gong
}
