% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vplot.r
\name{vplot_core}
\alias{vplot_core}
\title{vplot_core}
\usage{
vplot_core(x, field, ...)
}
\arguments{
\item{x}{a Vplots object}

\item{field}{The assays field that is used for Vplots (default: 'counts')}

\item{...}{Arguments passed to image()}
}
\description{
Plot a Vplot object
}
